flail

+++ lindenmayer/src/png.rs
@@ -17,7 +17,7 @@ fn convert_svg_to_png(data: &[u8]) -> Pixmap {
     let rtree = usvg::Tree::from_data(data, &usvg::Options::default()).unwrap();
 
     let pixmap_size = rtree.svg_node().size.to_screen_size();
-    let mut pixmap =  Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();
+    let mut pixmap =  Pixmap::new(pixmap_size.width() as u32, pixmap_size.height() as u32).unwrap();
     resvg::render(&rtree, FitTo::Original, pixmap.as_mut()).unwrap();
 
     pixmap
